integrations/sie_haystack/pyproject.toml
[project]
name = "sie-haystack"
version = "0.1.7"
description = "SIE integration for Haystack"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Superlinked", email = "dev@superlinked.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]

dependencies = [
    "sie-sdk>=0.1.0",
    "haystack-ai>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "chromadb>=0.4.0",
    "chroma-haystack>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sie_haystack"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-v --tb=short -m 'not integration'"
markers = ["integration: marks tests as integration tests (require running server)"]
